+++ src/main/java/com/Isabela01vSilva/bank_isabela/controller/response/account/UpdateAccountStatusResponse.java
@@ -3,6 +3,7 @@ package com.Isabela01vSilva.bank_isabela.controller.response.account;
 import com.Isabela01vSilva.bank_isabela.domain.account.AccountStatus;
 import com.Isabela01vSilva.bank_isabela.domain.account.AccountType;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record UpdateAccountStatusResponse(
@@ -10,7 +11,7 @@ public record UpdateAccountStatusResponse(
         String agencyNumber,
         AccountType accountType,
         AccountStatus accountStatus,
-        Double balance,
+        BigDecimal balance,
         LocalDate creationDate,
         LocalDate statusChangeDate,
         String statusChangeReason
